refactor: centralize MIMO layout ownership
