+++ flagscale/models/megatron/qwen35/__init__.py
@@ -5,7 +5,7 @@ from .attention import Qwen35SelfAttention
 from .language_model import Qwen35LanguageModule
 from .layer_specs import get_qwen35_language_model_spec
 from .qwen35_model import Qwen35Model
-from .qwen35_mimo_model import Qwen35MIMOModel
+from .qwen35_mimo_model import build_qwen35_mimo_model
 from .rope import Qwen35LanguageRotaryEmbedding, get_rope_index
 from .transformer_config import (
     Qwen35TransformerConfig,
@@ -18,7 +18,7 @@ from flagscale.models.megatron.qwen3_vl.layer_specs import get_qwen3vl_vision_mo
 
 __all__ = [
     "Qwen35Model",
-    "Qwen35MIMOModel",
+    "build_qwen35_mimo_model",
     "Qwen35LanguageModule",
     "Qwen35TransformerConfig",
     "get_vision_model_config",
