--- flagscale/models/mimo/bridge/recipe/__init__.py
@@ -1,3 +1,0 @@
-# Copyright (c) 2026, BAAI. All rights reserved.
-
-"""Model-specific grid training recipes (layout contracts, fail-fast rules)."""
