+++ flagscale/models/mimo/colocated/providers/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (c) 2026, BAAI. All rights reserved.
+
+"""Model providers for the colocated path (symmetric with ``grid.providers``)."""
